+++ relay-metrics/src/meta/aggregator.rs
@@ -3,106 +3,10 @@ use std::{
     hash::Hash,
 };
 
-use chrono::{DateTime, Utc};
 use relay_base_schema::project::ProjectKey;
-use relay_common::time::UnixTimestamp;
-use serde::{Deserialize, Serialize};
 
+use super::{Item, Location, MetricMeta, StartOfDayUnixTimestamp};
 use crate::{statsd::MetricCounters, MetricResourceIdentifier};
-
-/// A metric metadata item.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct MetricMeta {
-    /// Timestamp scope for the contained metadata.
-    ///
-    /// Metric metadata is collected in daily intervals, so this may be truncated
-    /// to the start of the day (UTC) already.
-    pub timestamp: StartOfDayUnixTimestamp,
-
-    /// The contained metadata mapped by MRI.
-    pub mapping: HashMap<MetricResourceIdentifier<'static>, Vec<MetricMetaItem>>,
-}
-
-/// A metadata item.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[serde(tag = "type", rename_all = "lowercase")]
-pub enum MetricMetaItem {
-    /// A location metadata pointing to the code location where the metric originates from.
-    Location(Location),
-    /// Unknown item.
-    #[serde(other)]
-    Unknown,
-}
-
-/// A code location.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, Deserialize, Serialize)]
-pub struct Location {
-    /// The relative file path.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub filename: Option<String>,
-    /// The absolute file path.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub abs_path: Option<String>,
-    /// The containing module name or path.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub module: Option<String>,
-    /// The containing function name.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub function: Option<String>,
-    /// The line number.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lineno: Option<u64>, // TODO nachschauen
-}
-
-/// A Unix timestamp that is truncated to the start of the day.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct StartOfDayUnixTimestamp(UnixTimestamp);
-
-impl StartOfDayUnixTimestamp {
-    /// Creates a new `StartOfDayUnixTimestamp` from a timestamp by truncating it.
-    ///
-    /// May return none when passed an invalid date, but in practice this never fails
-    /// since the [`UnixTimestamp`] is already sufficiently validated.
-    pub fn new(ts: UnixTimestamp) -> Option<Self> {
-        let dt: DateTime<Utc> = DateTime::from_timestamp(ts.as_secs().try_into().ok()?, 0)?;
-        let beginning_of_day = dt.date_naive().and_hms_opt(0, 0, 0)?.and_utc();
-        Some(Self(UnixTimestamp::from_datetime(beginning_of_day)?))
-    }
-
-    /// Returns the underlying unix timestamp, truncated to the start of the day.
-    pub fn as_timestamp(&self) -> UnixTimestamp {
-        self.0
-    }
-}
-
-impl std::ops::Deref for StartOfDayUnixTimestamp {
-    type Target = UnixTimestamp;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Serialize for StartOfDayUnixTimestamp {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        self.0.serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for StartOfDayUnixTimestamp {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let ts = UnixTimestamp::deserialize(deserializer)?;
-
-        StartOfDayUnixTimestamp::new(ts)
-            .ok_or_else(|| serde::de::Error::custom("invalid timestamp"))
-    }
-}
 
 /// A metrics meta aggregator.
 ///
@@ -186,7 +90,7 @@ impl MetaAggregator {
                 .mapping
                 .entry(scope.mri.clone()) // This clone sucks
                 .or_insert_with(Vec::new)
-                .extend(locations.iter().cloned().map(MetricMetaItem::Location));
+                .extend(locations.iter().cloned().map(Item::Location));
         }
 
         result.into_values()
@@ -198,28 +102,24 @@ impl MetaAggregator {
             .retain(|scope, _| scope.project_key != project_key);
     }
 
-    fn add_scoped(
-        &mut self,
-        scope: &Scope,
-        items: Vec<MetricMetaItem>,
-    ) -> Option<Vec<MetricMetaItem>> {
+    fn add_scoped(&mut self, scope: &Scope, items: Vec<Item>) -> Option<Vec<Item>> {
         // Entry ref needs hashbrown, we would have to clone the scope without or do a separate lookup.
         let locations = self.locations.entry_ref(scope).or_default();
         let mut send_upstream = Vec::new();
 
         for item in items {
             match item {
-                MetricMetaItem::Location(location) => {
+                Item::Location(location) => {
                     if locations.len() > self.max_locations {
                         break;
                     }
 
                     if !locations.contains(&location) {
                         locations.insert(location.clone());
-                        send_upstream.push(MetricMetaItem::Location(location));
+                        send_upstream.push(Item::Location(location));
                     }
                 }
-                MetricMetaItem::Unknown => {}
+                Item::Unknown => {}
             }
         }
